pjdfstest::context

Struct SerializedTestContext

pub struct SerializedTestContext<'a> {
    ctx: TestContext<'a>,
}

Serialized test context which allows to execute functions as another user/group(s) and with another umask.

Example

use nix::sys::stat::Mode;
use tempfile::TempDir;

use crate::{
  config::{Config, DummyAuthEntry},
  context::{FileType, SerializedTestContext},
};

let config = Config::default();
let temp_dir = TempDir::new().unwrap();
let entries = vec![DummyAuthEntry::new("user", "group")];
let ctx = SerializedTestContext::new(&config, &entries, temp_dir.path());

let (user, group) = ctx.get_new_entry();
assert_eq!(user.name, "user");
assert_eq!(group.name, "group");

ctx.as_user(user, None, || {
   let file = ctx.create(FileType::Regular).unwrap();
   let file_stat = nix::sys::stat::stat(&file).unwrap();
   assert_eq!(file_stat.st_mode & Mode::S_IFMT.bits(), nix::libc::S_IFREG);
});

Fields
ctx: TestContext<'a>
Implementations

impl<'a> SerializedTestContext<'a>

pub fn new( config: &'a Config, entries: &'a [DummyAuthEntry], base_dir: &'a Path, ) -> Self

pub fn as_user<F>(&self, user: &User, groups: Option<&[Gid]>, f: F)
where F: FnOnce(),

Execute the function as another user/group(s). If groups is set to None, only the default group associated to the user will be used and the effective [Gid] will be this one. Otherwise, the first provided [Gid] will be the effective one and the others will be added with setgroups.

pub fn with_umask<F>(&self, mask: mode_t, f: F)
where F: FnOnce(),

Execute the function with another umask.