pjdfstest::context

Struct TestContext

pub struct TestContext<'a> {
    naptime: Duration,
    temp_dir: &'a Path,
    features_config: &'a FeaturesConfig,
    auth_entries: DummyAuthEntries<'a>,
}

Test context which allows to create files, directories, get auth entries, and sleep for the amount of time specified in the configuration.

Note

The context will automatically remove all files and directories created during the test.

Example

use nix::sys::stat::Mode;
use tempfile::TempDir;

use crate::{
   config::{Config, DummyAuthEntry},
   context::{FileType, TestContext},
};

let config = Config::default();
let temp_dir = TempDir::new().unwrap();
let entries = vec![DummyAuthEntry::new("user", "group")];
let ctx = TestContext::new(&config, &entries, temp_dir.path());

let (user, group) = ctx.get_new_entry();
assert_eq!(user.name, "user");
assert_eq!(group.name, "group");

let file = ctx.create(FileType::Regular).unwrap();
let file_stat = nix::sys::stat::stat(&file).unwrap();
assert_eq!(file_stat.st_mode & Mode::S_IFMT.bits(), nix::libc::S_IFREG);

Fields
naptime: Duration

Duration to sleep, used to wait for file system timestamps to change.

temp_dir: &'a Path

Temporary directory where the test will be executed and files will be created.

features_config: &'a FeaturesConfig

Features configuration, used to determine which features are enabled.

auth_entries: DummyAuthEntries<'a>

Auth entries which are composed of a [User] and its associated [Group].

Implementations

impl<'a> TestContext<'a>

pub fn new( config: &'a Config, entries: &'a [DummyAuthEntry], temp_dir: &'a Path, ) -> Self

Create a new test context.

pub fn base_path(&self) -> &Path

Return the base path for this context.

pub fn features_config(&self) -> &FeaturesConfig

pub fn gen_path(&self) -> PathBuf

Generate a random path.

pub fn create_file( &self, oflag: OFlag, mode: Option<mode_t>, ) -> Result<(PathBuf, OwnedFd), Error>

Create a regular file and open it.

pub fn new_file(&self, ft: FileType) -> FileBuilder

Return a file builder.

pub fn create(&self, f_type: FileType) -> Result<PathBuf, Error>

Create a file with a random name.

pub fn create_name_max(&self, f_type: FileType) -> Result<PathBuf, Error>

Create a file whose name length is _PC_NAME_MAX.

pub fn create_path_max(&self, f_type: FileType) -> Result<PathBuf, Error>

Create a file whose path length is _PC_PATH_MAX.

pub fn get_new_entry(&self) -> (&User, &Group)

Returns a new entry.

pub fn get_new_user(&self) -> &User

Returns a new user. Alias of get_new_entry.

pub fn get_new_group(&self) -> &Group

Returns a new group. Alias of get_new_entry.

pub fn nap(&self)

A short sleep, long enough for file system timestamps to change.