pjdfstest

Module test

Test framework for testing the filesystem implementation.

Re-exports

pub use crate::context::SerializedTestContext;

pub use crate::context::TestContext;

pub use crate::features::*;

pub use crate::flags::*;

Structs

TestCase

A single minimal test case.

Enums

TestFn

Function which runs the test. The function is passed a context object which can be used to interact with the filesystem.

Type Aliases

Guard

Function which indicates if the test should be skipped by returning an error.