pjdfstest

Module tests

Where the tests are defined.

This module contains the tests for the file system test suite. The tests are defined as functions which are then registered with the test suite.

The tests are defined using the test_case! macro, which is used to define a test case for the test suite.

It also contains some helper functions and macros which are used to define the tests.

Modules

chmod

chown

errors

Helper functions for testing error handling.

ftruncate

link

mkdir

mkfifo

mknod

mksyscalls 🔒

Builder functions for mk-family syscalls tests.

open

posix_fallocate

rename

rmdir

symlink

truncate

unlink

utimensat

Structs

InvariantTimeMetadata 🔒

Metadata which isn’t related to time.

TimeAssertion 🔒

Builder to create a time metadata assertion, which compares metadata between pairs of paths.

TimestampField 🔒

Argument to set which fields should be compared for TimeAssertion::path.

Constants

ATIME 🔒

CTIME 🔒

MTIME 🔒

Traits

AsTimeInvariant 🔒

MetadataExt 🔒

A handy extention to std::os::unix::fs::MetadataExt

Functions

assert_ctime_changed 🔒

Assert that a certain operation changes the ctime of a file.

assert_ctime_unchanged 🔒

Assert that a certain operation does not change the ctime of a file.

assert_mtime_changed 🔒

Assert that a certain operation changes the mtime of a file.

assert_symlink_ctime_unchanged 🔒

Assert that a certain operation does not change the ctime of a file without following symlinks.

assert_times_changed 🔒

Alias for TimeAssertion::new(false).

assert_times_unchanged 🔒

Alias for TimeAssertion::new(true).