pjdfstest::tests::errors

Module eloop

Macros

eloop_comp_test_case 🔒

Create a test case which asserts that the sycall returns ELOOP if too many symbolic links were encountered in translating a component of the pathname which is not the last one. There are multiple forms for this macro:

eloop_either_test_case 🔒

Create a test case which asserts that the sycall returns ELOOP if too many symbolic links were encountered in translating a component of either pathname which is not the last one.

eloop_final_comp_test_case 🔒

Create a test case which asserts that the sycall returns ELOOP if too many symbolic links were encountered in translating the last component of the pathname. There are multiple forms for this macro:

Functions

create_loop_symlinks

Create a loop between two symbolic links and return them.