Macro pjdfstest::tests::errors::etxtbsy::etxtbsy_test_case

macro_rules! etxtbsy_test_case {
    ($syscall: ident, $($f: expr),+) => { ... };
    ($syscall: ident $( ($( $($before:expr),* ,)? ~path $(, $($after:expr),*)?) )?) => { ... };
}

Create a test case which asserts that it returns ETXTBSY when the file is a pure procedure (shared text) file that is being executed. There are multiple forms for this macro:

A basic form which takes the syscall, and optionally a ~path argument to indicate where the path argument should be substituted if the path is not the only argument taken by the syscall.

// `unlink` accepts only a path as argument.
etxtbsy_test_case!(unlink);
// `truncate` takes a path and the flags to set as arguments.
// We need to add `~path` where the path argument should normally be taken.
etxtbsy_test_case!(truncate(~path, 123));

A more complex form which takes multiple functions with the context and the path as arguments for syscalls requring to compute other arguments.

etxtbsy_test_case!(chown, |ctx: &mut TestContext, path: &Path| {
  let user = ctx.get_new_user();
  chown(path, Some(user.uid), None)
})