pjdfstest::tests

Module mksyscalls

Builder functions for mk-family syscalls tests.

Functions

assert_perms_from_mode_and_umask 🔒

Assert that the created entry gets its permission bits from the mode provided to the function negated by the process’s file creation mask (umask), and its file type is equal to the expected one.

assert_uid_gid 🔒

Assert that the entry’s user ID is set to the process’ effective user ID and the entry’s group ID should be set to the group ID of the parent directory or the effective group ID of the process.