pjdfstest

Module utils

Utility functions for filesystem operations.

This module provides utility functions for filesystem operations which are not available in the standard library.

Modules

dev

Constants

ALLPERMS

Functions

chmod

Wrapper for fchmodat(None, path, mode, FchmodatFlags::FollowSymlink).

get_mountpoint

Get mountpoint.

lchmod

Wrapper for fchmodat(None, path, mode, FchmodatFlags::NoFollowSymlink).

lchown

Wrapper for fchownat(None, path, mode, FchownatFlags::NoFollowSymlink).

link

Wrapper for linkat(None, old_path, None, new_path).

open

Wrapper for open which returns Ownedfd instead of RawFd.

rename

Wrapper for renameat(None, old_path, None, new_path).

rmdir

Wrapper for rmdir.

symlink

Wrapper for symlinkat(path1, None, path2).