Crate pjdfstest

This is the main entry point for the test suite. It is responsible for parsing command line arguments, reading the configuration file, and running the tests. The test suite is composed of a set of test cases, each of which is a function that takes a TestContext as an argument. The TestContext provides access to the configuration, the dummy authentication entries, and a temporary directory for the test to use.

The test suite is built using the inventory crate, which allows test cases to be registered at compile time. The test suite is run by iterating over the registered test cases and running each one in turn.

The TestContext is created for each test case, and the test case function is called with the TestContext as an argument. The test case function can then use the TestContext to access the configuration, the dummy authentication entries, and the temporary directory. The test case function can perform whatever tests are necessary, and panic if the test fails. The test suite catches the panic, prints an error message, and continues running the remaining test cases. At the end of the test suite, the number of failed, skipped, and passed tests is printed.

Modules

config 🔒

Configuration for the test suite.

context 🔒

Provides a context for tests through the TestContext and SerializedTestContext structs.

features 🔒

File-system features which are not available on every file system and can be tested for.

flags 🔒

File flags (see https://docs.freebsd.org/en/books/handbook/basics/#permissions).

macros 🔒

Macros for defining test cases.

test 🔒

Test framework for testing the filesystem implementation.

tests 🔒

Where the tests are defined.

utils 🔒

Utility functions for filesystem operations.

Structs

ArgOptions 🔒

Statics

BACKTRACE 🔒

Functions

main 🔒

run_test_cases 🔒

Run provided test cases and filter according to features and flags availability.