Module pjdfstest::context

Provides a context for tests through the TestContext and SerializedTestContext structs.

The TestContext struct allows to create files, directories, get auth entries, and sleep for the amount of time specified in the configuration. Its SerializedTestContext counterpart allows to execute functions as another user/group(s) and with another umask.

Structs

DummyAuthEntries

Auth entries which are composed of a [User] and its associated [Group]. Allows to retrieve the auth entries.

FileBuilder

Allows to create a file using builder pattern.

FileTypeIter

An iterator over the variants of Self

SerializedTestContext

Serialized test context which allows to execute functions as another user/group(s) and with another umask.

TestContext

Test context which allows to create files, directories, get auth entries, and sleep for the amount of time specified in the configuration.

Enums

FileType

File type, mainly used with TestContext::create and parameterized tests.

Constants

NUM_RAND_CHARS 🔒