Macro pjdfstest::macros::test_case

macro_rules! test_case {
    ($(#[doc = $docs:expr])*
        $f:ident, serialized, root $(,)* $( $features:expr ),* $(,)* $(; $( $flags:expr ),+)? $(=> $guards: tt )?) => { ... };
    ($(#[doc = $docs:expr])*
        $f:ident, serialized $(,)* $( $features:expr ),* $(,)* $(; $( $flags:expr ),+)? $(=> $guards: tt )?) => { ... };
    ($(#[doc = $docs:expr])*
        $f:ident, root $(,)* $( $features:expr ),* $(,)* $(; $( $flags:expr ),+)? $(=> $guards: tt )?) => { ... };
    ($(#[doc = $docs:expr])*
        $f:ident $(,)* $( $features:expr ),* $(,)* $(; $( $flags:expr ),+)? $(=> $guards: tt )?) => { ... };
    (@serialized $f:ident, $features:expr, $guards:expr, $desc:expr, $require_root:expr ) => { ... };
    (@serialized $f:ident, $features:expr, $guards:expr, $desc:expr, $require_root:expr => [$( $file_type:tt $( ($ft_args: tt) )? ),+ $(,)*]) => { ... };
    (@ $f:ident, $features:expr, $guards:expr, $require_root:expr, $desc:expr ) => { ... };
    (@ $f:ident, $features:expr, $guards:expr, $require_root:expr, $desc:expr => [$( $file_type:tt $( ($ft_args: tt) )? ),+ $(,)*]) => { ... };
}

Macro for defining test cases, which are automatically registered with the test suite.

A test case can be serialized or non-serialized, require root privileges, and be run on specific file types. It can also require specific features to be enabled, and have guards which are run before the test case is executed to determine if conditions are met.

The macro supports mutiple parameters which can be combined in a specific order, for example:

// Non-serialized test case
test_case! {
   /// description
  basic
}
fn basic(_: &mut crate::test::TestContext) {}

// Non-serialized test case with required features, guards, and root privileges
test_case! {
  /// description
 features, root, FileSystemFeature::Chflags, FileSystemFeature::PosixFallocate; guard_example
}
fn features(_: &mut crate::test::TestContext) {}

// Serialized test case with root privileges
test_case! {
 /// description
serialized, serialized, root
}
fn serialized(_: &mut crate::test::SerializedTestContext) {}

// Serialized test case with required features
test_case! {
/// description
serialized_features, serialized, FileSystemFeature::Chflags, FileSystemFeature::PosixFallocate
}
fn serialized_features(_: &mut crate::test::SerializedTestContext) {}

// Serialized test case with required features, guards, root privileges, and file types
test_case! {
/// description
serialized_types, serialized, FileSystemFeature::Chflags, FileSystemFeature::PosixFallocate; guard_example, root => [Regular, Fifo]
}
fn serialized_types(_: &mut crate::test::SerializedTestContext, _: crate::context::FileType) {}