Function pjdfstest::tests::mksyscalls::assert_perms_from_mode_and_umask

pub(super) fn assert_perms_from_mode_and_umask<F, T, C>(
    ctx: &mut SerializedTestContext<'_>,
    f: F,
    f_type_check: C,
)
where
    F: Fn(&Path, Mode) -> Result<T>,
    C: Fn(&FileType) -> bool,

Assert that the created entry gets its permission bits from the mode provided to the function negated by the process’s file creation mask (umask), and its file type is equal to the expected one.